Split waveform.py into a sub-package
File	Responsibility	Lines
waveform/units.py	ELL_MIN, ELL_MAX, _modal_dt	~18
waveform/loaders.py	load_from_h5, load_from_targz as standalone functions	~210
waveform/matching.py	apply_wigner_rotation_to_mode_dict, interpolate_in_amp_phase	~100
waveform/modes.py	WaveformModes class (all instance methods; classmethods delegate to loaders)	~500
waveform/__init__.py	Re-exports full public API	~20
